+++ CHANGELOG.md
@@ -57,6 +57,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
+- Drop `AbstractModel::comesFromDatabase()` (#3969)
 - Drop automatic salutations from the backend module emails (#3959)
 - Drop the hooks for the email form in the BE module (#3954)
 - Drop code for paid/unpaid registrations (#3950)